+++ libs/template/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [16.0.2](https://github.com/rx-angular/rx-angular/compare/template@16.0.1...template@16.0.2) (2023-07-17)
+
+
+### Bug Fixes
+
+* **template:** autosize virtual scroll handles updates with trackBy ([5d1034e](https://github.com/rx-angular/rx-angular/commit/5d1034e3a6526f684308bd29fb48dbc21f6daade))
+* **template:** virtual-scroll scrollStrategy dependency error in es2022 ([#1586](https://github.com/rx-angular/rx-angular/issues/1586)) ([45ead14](https://github.com/rx-angular/rx-angular/commit/45ead1435e1a1d881937a2d913252a9f38ed06dc))
+* **template:** virtual-scrolling rxjs 6 compat. autosize update handling ([57588ea](https://github.com/rx-angular/rx-angular/commit/57588ea236baed54086e1a3b9db340722bf7bde6))
+
+
+
 ## [16.0.1](https://github.com/rx-angular/rx-angular/compare/template@16.0.0...template@16.0.1) (2023-06-22)